Bootstrap: Center Text with Bootstrap
<!-- Bootstrap: Center Text with Bootstrap 
Now that we're using Bootstrap, we can center our heading element to make it look better. All we need to do is add the class text-center to our h2 element.

Remember that you can add several classes to the same element by separating each of them with a space, like this:

<h2 class="red-text text-center">your text</h2> -->