Bootstrap: Warn Your Users of a Dangerous Action with btn-danger
<!-- Bootstrap: Warn Your Users of a Dangerous Action with btn-danger
Bootstrap comes with several pre-defined colors for buttons. The btn-danger class is the button color you'll use to notify users that the button performs a destructive action, such as deleting a cat photo.

Create a button with the text "Delete" and give it the class btn-danger.
Note that these buttons still need the btn and btn-block classes.
-->